1、修复直播、点播暂停后，退出后台回到界面，会自动播放的问题 2、修复超级播放器添加视频后，列表中不出现视频的问题 3、pub版本号修改 4、修复切换清晰度列表后，列表不消失的问题 5、画中画Controller，增加画中画退出回调，用户可自行控制画中画退出后的操作 6、播放视频增加版本log输出 7、修复安卓画中画不能播放加密视频的问题 (merge request !127)
Squash merge branch 'kongdywang/main/fix_pip_encrypt_error' into 'main'
1、修复直播、点播暂停后，退出后台回到界面，会自动播放的问题
2、修复超级播放器添加视频后，列表中不出现视频的问题
3、pub版本号修改
4、修复切换清晰度列表后，列表不消失的问题
5、画中画Controller，增加画中画退出回调，用户可自行控制画中画退出后的操作
6、播放视频增加版本log输出
7、修复安卓画中画不能播放加密视频的问题
8、修复退出画中画之后，有安卓异常log输出的问题

+++ Flutter/ios/Classes/FTXAudioManager.h
@@ -6,9 +6,9 @@
 @interface FTXAudioManager : NSObject
 
 
-- (float)getVolume;
+- (CGFloat)getVolume;
 
-- (void)setVolume:(float)value;
+- (void)setVolume:(CGFloat)value;
 
 - (void)setVolumeUIVisible:(BOOL)volumeUIVisible;
